browser/devtools/webide/test/test_autoconnect_runtime.html

<!DOCTYPE html>

<html>

  <head>
    <meta charset="utf8">
    <title></title>

    <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript" src="chrome://mochikit/content/chrome-harness.js"></script>
    <script type="application/javascript;version=1.8" src="head.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  </head>

  <body>

    <script type="application/javascript;version=1.8">
      window.onload = function() {
        SimpleTest.waitForExplicitFinish();

        Task.spawn(function* () {

          Cu.import("resource://gre/modules/devtools/dbg-server.jsm");

          if (!DebuggerServer.initialized) {
            DebuggerServer.init();
            DebuggerServer.addBrowserActors();
          }

          let win = yield openWebIDE();

          let fakeRuntime = {
            type: "USB",
            connect: function(connection) {
              is(connection, win.AppManager.connection, "connection is valid");
              connection.host = null; // force connectPipe
              connection.connect();
              return promise.resolve();
            },

            get id() {
              return "fakeRuntime";
            },

            get name() {
              return "fakeRuntime";
            }
          };
          win.AppManager.runtimeList.usb.push(fakeRuntime);
          win.AppManager.update("runtime-list");

          let panelNode = win.document.querySelector("#runtime-panel");
          let items = panelNode.querySelectorAll(".runtime-panel-item-usb");
          is(items.length, 1, "Found one runtime button");

          let deferred = promise.defer();
          win.AppManager.connection.once(
              win.Connection.Events.CONNECTED,
              () => deferred.resolve());
          items[0].click();

          ok(win.document.querySelector("window").className, "busy", "UI is busy");
          yield win.UI._busyPromise;
          is(Object.keys(DebuggerServer._connections).length, 1, "Connected");

          yield nextTick();

          yield closeWebIDE(win);

          is(Object.keys(DebuggerServer._connections).length, 0, "Disconnected");

          win = yield openWebIDE();

          win.AppManager.runtimeList.usb.push(fakeRuntime);
          win.AppManager.update("runtime-list");

          yield waitForUpdate(win, "runtime-targets");

          is(Object.keys(DebuggerServer._connections).length, 1, "Automatically reconnected");

          yield win.Cmds.disconnectRuntime();

          yield closeWebIDE(win);

          DebuggerServer.destroy();

          SimpleTest.finish();
        });
      }


    </script>
  </body>
</html>
